Added dagger
The dagger is a short range weapon. It has the same dash mechanic as the lance, but with a twist. The dagger activates 3 times with each trigger pull and with each activation the player can aim the dagger in a new direction. The player gets a velocity boost in whichever new direction they have chosen for each activation.